requirements.txt
aiofiles==22.1.0
aiosqlite==0.17.0
anyio==3.6.1
asyncpg==0.27.0
bcrypt==3.2.2
cffi==1.15.1
click==8.1.3
colorama==0.4.5
databases==0.6.0
fastapi==0.85.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
passlib==1.7.4
Pillow==9.2.0
pycparser==2.21
pydantic==1.9.1
python-dotenv==0.21.0
python-multipart==0.0.5
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.40
starlette==0.20.4
typing_extensions==4.3.0
uvicorn==0.17.6
